requirements.txt
black==21.5b1
coveralls==3.1.0
discord==1.0.1
fuzzywuzzy==0.18.0
mouseinfo==0.1.3
nest-asyncio==1.5.1
pip-chill==1.0.1
pre-commit==2.13.0
pygetwindow==0.0.9
pymsgbox==1.0.9
pyscreeze==0.1.27
python-levenshtein==0.12.2
pytweening==1.0.3
sphinx-rtd-theme==0.5.2
twine==3.4.1

Sphinx~=3.5.4
attrs~=20.3.0
setuptools~=56.0.0